`uintptr_t` is now used for address casting.

+++ src/LiquidLine.cpp
@@ -26,7 +26,7 @@ SOFTWARE.
 
 //const char NOTHING[] = "                    "; ///< 20 spaces, used to clear the display.
 
-void print_me(uint16_t address) {
+void print_me(uintptr_t address) {
 	DEBUG(F("Line (0x")); DEBUG2(address, OCT); DEBUG(F("): "));
 	return;
 	address = address;
@@ -34,7 +34,7 @@ void print_me(uint16_t address) {
 
 
 bool LiquidLine::attach_function(uint8_t number, void (*function)(void)) {
-	print_me((uint16_t)this);
+	print_me((uintptr_t)this);
 	if (number <= MAX_FUNCTIONS) {
 		_function[number - 1] = function;
 		DEBUG(F("Attached function ")); DEBUGLN(number);
@@ -48,7 +48,7 @@ bool LiquidLine::attach_function(uint8_t number, void (*function)(void)) {
 }
 
 bool LiquidLine::set_focusPosition(Position position, uint8_t column, uint8_t row) {
-	print_me((uint16_t)this);
+	print_me((uintptr_t)this);
 	if (position <= Position::CUSTOM) {
 		_focusPosition = position;
 		_focusColumn = column;
