+++ src/LiquidMenu.cpp
@@ -63,7 +63,7 @@ LiquidMenu::LiquidMenu(LiquidCrystal &liquidCrystal, LiquidScreen &liquidScreen1
 }
 
 bool LiquidMenu::add_screen(LiquidScreen &liquidScreen) {
-	print_me((uint16_t)this);
+	print_me((uintptr_t)this);
 	if (_screenCount < MAX_SCREENS) {
 		_p_liquidScreen[_screenCount] = &liquidScreen;
 		DEBUG(F("Added a new screen (")); DEBUG(_screenCount); DEBUGLN(F(")"));
@@ -136,7 +136,7 @@ bool LiquidMenu::change_screen(LiquidScreen &p_liquidScreen) {
 	// _p_liquidMenu[_currentMenu]->_p_liquidCrystal->clear();
 	for (uint8_t s = 0; s < _screenCount; s++) {
 		_p_liquidCrystal->clear();
-		if ((uint16_t)&p_liquidScreen == (uint16_t) & (*_p_liquidScreen[s])) {
+		if ((uintptr_t)&p_liquidScreen == (uintptr_t) & (*_p_liquidScreen[s])) {
 			_currentScreen = s;
 			update();
 			DEBUG(F("Switched to screen ("));
@@ -145,7 +145,7 @@ bool LiquidMenu::change_screen(LiquidScreen &p_liquidScreen) {
 			return true;
 		}
 	}
-	DEBUG(F("Invalid request for screen change to ")); DEBUGLN((uint16_t)&p_liquidScreen);
+	DEBUG(F("Invalid request for screen change to ")); DEBUGLN((uintptr_t)&p_liquidScreen);
 	return false;
 }
 
@@ -164,7 +164,7 @@ void LiquidMenu::switch_focus(bool forward) {
 }
 
 bool LiquidMenu::set_focusPosition(Position position) {
-	print_me((uint16_t)this);
+	print_me((uintptr_t)this);
 	if (position == Position::CUSTOM) {
 		DEBUGLN(F("Can't set a custom focus position for the whole menu at once"));
 		return false;
